+++ database/migrations/2025_06_13_030328_create_claim_uploads_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -16,9 +15,9 @@ return new class extends Migration
             $table->string('customer_name')->nullable();
             $table->integer('sheet_value')->nullable();
             $table->integer('recipe_value')->nullable();
-            $table->integer('commercial_value')->nullable();
-            $table->integer('tax_value')->nullable();
-            $table->integer('total')->nullable();
+            $table->bigInteger('commercial_value')->nullable();
+            $table->bigInteger('tax_value')->nullable();
+            $table->bigInteger('total')->nullable();
             $table->foreignId('unitbisnis_code')->nullable();
             $table->foreignId('user_id')->nullable();
             $table->string('period')->nullable();
